docker-airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
# pyspark
sqlalchemy
psycopg2-binary